+++ CHANGELOG.md
@@ -14,6 +14,9 @@ the [PEP 440 version scheme](https://peps.python.org/pep-0440/#version-scheme).
 ### Fixed
 ### Removed
 ### Security
+- A bug in pydantic(https://github.com/pydantic/pydantic/issues/9139) causes
+  `SecretString` to still be logged plaintext in case of a validation error.
+  This bug is circumvented by hiding all input in validation errors by default.
 
 
 ## [2.4.0] - 2026-08-19